+++ database/migrations/2021_08_03_115452_add_object_type_to_item_translations_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddObjectTypeToItemTranslationsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('item_translations', function (Blueprint $table) {
+            $table->string('object_type')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('item_translations', function (Blueprint $table) {
+            $table->dropColumn('object_type');
+        });
+    }
+}