+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Added
 - Typeform integration
+- item object type attribute
 
 ## [2.29.2] - 2021-06-11
 ### Added